change partiality annotation and add a lie in allocation
We want a better crash message if allocation fails but we also want totality.
It's morally correct to treat allocation of bytes the same as allocation
of space for :: Both are total, bytes allocation just happens to be more
likely to crash from running out of resources because the required space is
dynamicly determined. This is why we want a more specific error message
for it.  We shouldn't have to sacrifice by peppering the whole codebase
with 'partial' just to have a nicer error message on an allocation crash.